Rename / refactor
rename & refactor a bit before LotATC work
