+++ DCS-SR-Client/DCS-SR-Client.csproj
@@ -190,10 +190,12 @@
     <Compile Include="Audio\Providers\VolumeSampleProviderWithPeak.cs" />
     <Compile Include="Audio\Utility\SpeexWrapper.cs" />
     <Compile Include="Input\InputBindState.cs" />
-    <Compile Include="Network\DCSAutoConnectListener.cs" />
-    <Compile Include="Network\Models\CombinedRadioState.cs" />
-    <Compile Include="Network\Models\DCSLosCheckRequest.cs" />
-    <Compile Include="Network\Models\DCSLosCheckResult.cs" />
+    <Compile Include="Network\DCS\DCSAutoConnectHandler.cs" />
+    <Compile Include="Network\DCS\DCSLineOfSightHandler.cs" />
+    <Compile Include="Network\DCS\Models\CombinedRadioState.cs" />
+    <Compile Include="Network\DCS\Models\DCSLosCheckRequest.cs" />
+    <Compile Include="Network\DCS\Models\DCSLosCheckResult.cs" />
+    <Compile Include="Network\UDPCommandHandler.cs" />
     <Compile Include="Network\UDPVoiceHandler.cs" />
     <Compile Include="Settings\SynchedServerSettings.cs" />
     <Compile Include="Properties\Resources.Designer.cs">
@@ -249,7 +251,7 @@
       <DependentUpon>RadioOverlay.xaml</DependentUpon>
     </Compile>
     <Compile Include="Network\Models\RadioSendingState.cs" />
-    <Compile Include="Network\RadioDCSSyncServer.cs" />
+    <Compile Include="Network\DCS\DCSRadioSyncHandler.cs" />
     <Compile Include="UI\ServerSettingsWindow\ServerSettingsWindow.xaml.cs">
       <DependentUpon>ServerSettingsWindow.xaml</DependentUpon>
     </Compile>
@@ -272,7 +274,7 @@
       <DependentUpon>App.xaml</DependentUpon>
       <SubType>Code</SubType>
     </Compile>
-    <Compile Include="Network\ClientSync.cs" />
+    <Compile Include="Network\SRSClientSyncHandler.cs" />
     <Compile Include="UI\ClientWindow\MainWindow.xaml.cs">
       <DependentUpon>MainWindow.xaml</DependentUpon>
       <SubType>Code</SubType>
