+++ DCS-SR-Client/Network/DCS/Models/CombinedRadioState.cs
