+++ DCS-SR-Client/Network/DCS/Models/DCSLosCheckRequest.cs
