+++ DCS-SR-Client/Network/DCS/Models/DCSLosCheckResult.cs
